libblkid: add to blkid_cache reference to blkid_config
The current blkid_cache implementation have no clue blkid_config. We
read the config just to get path to the cache file, but nothing else
is possible. It makes it difficult to extend and for example
evaluate.c has to explicitly care about path to the cache etc.

This patch add to blkid_cache reference to blkid_config and it
improves cache related functions to read cache or config in arbitrary
order (from code point of view, practically we always read config
before cache).

If you already have cache and need config:
	 blkid_get_config(cache);

or if you have config and need cache:
	 blkid_get_cache_for_config(...);

All the changes are in private libblkid stuff, nothing extends or
modifies library API.

Signed-off-by: Karel Zak <kzak@redhat.com>